Improve package name
