+++ cli/src/test/kotlin/io/spine/protodata/cli/given/DefaultOptionsCounterView.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.param.given
+package io.spine.protodata.cli.given
 
 import io.spine.core.External
 import io.spine.core.Subscribe
