+++ cli/src/test/kotlin/io/spine/protodata/param/CliTest.kt
@@ -36,11 +36,11 @@ import io.spine.json.Json
 import io.spine.option.OptionsProto
 import io.spine.protobuf.AnyPacker
 import io.spine.protodata.cli.Run
-import io.spine.protodata.param.given.CustomOptionPlugin
-import io.spine.protodata.param.given.CustomOptionRenderer
-import io.spine.protodata.param.given.DefaultOptionsCounterPlugin
-import io.spine.protodata.param.given.DefaultOptionsCounterRenderer
-import io.spine.protodata.param.given.TestOptionProvider
+import io.spine.protodata.cli.given.CustomOptionPlugin
+import io.spine.protodata.cli.given.CustomOptionRenderer
+import io.spine.protodata.cli.given.DefaultOptionsCounterPlugin
+import io.spine.protodata.cli.given.DefaultOptionsCounterRenderer
+import io.spine.protodata.cli.given.TestOptionProvider
 import io.spine.protodata.cli.test.TestOptionsProto
 import io.spine.protodata.cli.test.TestProto
 import io.spine.protodata.test.ECHO_FILE
